Create Covid-19 Data exploration
The SQL query provided is a comprehensive analysis of COVID-19 data from various sources. It includes calculations for total cases, new cases, total deaths, and new deaths, as well as the percentage of population infected and vaccinated. The query also includes aggregations by continent and country to provide insights into regional and national trends. The results are presented in a clear and organized manner, making it easier to visualize and understand the data.